refactor package processor
