+++ drools-compiler/src/main/java/org/drools/compiler/builder/impl/TypeDefinition.java
@@ -18,7 +18,7 @@ package org.drools.compiler.builder.impl;
 import org.drools.drl.ast.descr.AbstractClassTypeDeclarationDescr;
 import org.drools.core.rule.TypeDeclaration;
 
-class TypeDefinition {
+public class TypeDefinition {
 
     final AbstractClassTypeDeclarationDescr typeDescr;
     final TypeDeclaration type;
