+++ drools-compiler/src/main/java/org/drools/compiler/builder/impl/processors/AbstractPackageProcessor.java
@@ -1,4 +1,4 @@
-package org.drools.compiler.builder.impl;
+package org.drools.compiler.builder.impl.processors;
 
 import org.drools.compiler.compiler.PackageRegistry;
 import org.drools.drl.ast.descr.PackageDescr;
@@ -7,7 +7,7 @@ import org.kie.internal.builder.KnowledgeBuilderResult;
 import java.util.ArrayList;
 import java.util.Collection;
 
-public abstract class AbstractPackageProcessor {
+public abstract class AbstractPackageProcessor implements Processor {
     protected final PackageRegistry pkgRegistry;
     protected final PackageDescr packageDescr;
     protected final Collection<KnowledgeBuilderResult> results;
