+++ drools-compiler/src/main/java/org/drools/compiler/builder/impl/processors/EntryPointDeclarationProcessor.java
@@ -1,0 +1,18 @@
+package org.drools.compiler.builder.impl.processors;
+
+import org.drools.compiler.compiler.PackageRegistry;
+import org.drools.drl.ast.descr.EntryPointDeclarationDescr;
+import org.drools.drl.ast.descr.PackageDescr;
+
+public class EntryPointDeclarationProcessor extends AbstractPackageProcessor {
+    public EntryPointDeclarationProcessor(PackageRegistry pkgRegistry, PackageDescr packageDescr) {
+        super(pkgRegistry, packageDescr);
+    }
+
+    public void process() {
+        for (EntryPointDeclarationDescr epDescr : packageDescr.getEntryPointDeclarations()) {
+            pkgRegistry.getPackage().addEntryPointId(epDescr.getEntryPointId());
+        }
+    }
+
+}
